Move coalib.misc.Annotations into coalib.settings
- Move typechain(in coalib.misc.Annotations) to coalib.settings as
  all type modules are in coalib.settings.
- Move the corresponding test moudule(tests.misc.AnnotationsTest) to
  tests.settings

Fixes https://github.com/coala/coala/issues/3049
