+++ coalib/settings/Annotations.py
